+++ src/__init__.py
@@ -1,3 +1,3 @@
 #!/usr/bin/python3
 
-__version__: str = "2.23.2"
+__version__: str = "2.24.0"